initial update (w/o @artblocks/contracts pkg update)
