+++ src/da-exp-lib-mapping.ts
@@ -1,13 +1,14 @@
 import {
-  AuctionMinimumLengthSecondsUpdated,
-  SetAuctionDetailsLin
-} from "../generated/ISharedDALin/ISharedMinterDALinV0";
+  AuctionMinHalfLifeSecondsUpdated,
+  SetAuctionDetailsExp
+} from "../generated/DAExpLib/DAExpLib";
 
-import { loadOrCreateMinterProjectAndConfigIfProject } from "./i-shared-minter-mapping";
+import { loadOrCreateMinterProjectAndConfigIfProject } from "./generic-minter-events-lib-mapping";
 
 import {
   loadOrCreateMinter,
-  updateProjectIfMinterConfigIsActive
+  updateProjectIfMinterConfigIsActive,
+  getTotalDAExpAuctionTime
 } from "./helpers";
 
 import {
@@ -22,19 +23,19 @@ import {
 // minter-level configuration events
 
 /**
- * handles the update of a Linear Dutch Auction's minimum auction length in seconds.
- * @param event The event containing the updated minimum auction length in seconds.
+ * handles the update of Exponential Dutch Auction's minimum price decay half life in seconds.
+ * @param event The event containing the updated minimum price decay half life in seconds.
  */
-export function handleAuctionMinimumLengthSecondsUpdated(
-  event: AuctionMinimumLengthSecondsUpdated
+export function handleAuctionMinHalfLifeSecondsUpdated(
+  event: AuctionMinHalfLifeSecondsUpdated
 ): void {
   // load minter
   const minter = loadOrCreateMinter(event.address, event.block.timestamp);
 
   // update minter extra minter details value
   setMinterExtraMinterDetailsValue(
-    "minimumAuctionLengthInSeconds",
-    event.params._minimumAuctionLengthSeconds,
+    "minimumHalfLifeInSeconds",
+    event.params.minimumPriceDecayHalfLifeSeconds,
     minter
   );
 
@@ -47,14 +48,14 @@ export function handleAuctionMinimumLengthSecondsUpdated(
 
 /**
  * Handles the event that updates the auction details for a project using an
- * linear Dutch auction minter.
- * @param event The event carrying the updated lin auction data.
+ * exponential Dutch auction minter.
+ * @param event The event carrying the updated exp auction data.
  */
-export function handleSetAuctionDetailsLin(event: SetAuctionDetailsLin): void {
+export function handleSetAuctionDetailsExp(event: SetAuctionDetailsExp): void {
   const minterProjectAndConfig = loadOrCreateMinterProjectAndConfigIfProject(
     event.address, // minter
-    event.params._coreContract,
-    event.params._projectId,
+    event.params.coreContract,
+    event.params.projectId,
     event.block.timestamp
   );
   if (!minterProjectAndConfig) {
@@ -65,21 +66,33 @@ export function handleSetAuctionDetailsLin(event: SetAuctionDetailsLin): void {
   const projectMinterConfig = minterProjectAndConfig.projectMinterConfiguration;
   // remove the auction details from the project minter config
 
-  projectMinterConfig.basePrice = event.params._basePrice;
+  projectMinterConfig.basePrice = event.params.basePrice;
   projectMinterConfig.priceIsConfigured = true;
   setProjectMinterConfigExtraMinterDetailsValue(
     "startPrice",
-    event.params._startPrice.toString(), // Price is likely to overflow js Number.MAX_SAFE_INTEGER so store as string
+    event.params.startPrice.toString(), // Price is likely to overflow js Number.MAX_SAFE_INTEGER so store as string
     projectMinterConfig
   );
   setProjectMinterConfigExtraMinterDetailsValue(
     "startTime",
-    event.params._auctionTimestampStart,
+    event.params.auctionTimestampStart,
     projectMinterConfig
   );
   setProjectMinterConfigExtraMinterDetailsValue(
-    "endTime",
-    event.params._auctionTimestampEnd,
+    "halfLifeSeconds",
+    event.params.priceDecayHalfLifeSeconds,
+    projectMinterConfig
+  );
+
+  // pre-calculate the approximate DA end time
+  const totalAuctionTime = getTotalDAExpAuctionTime(
+    event.params.startPrice,
+    event.params.basePrice,
+    event.params.priceDecayHalfLifeSeconds
+  );
+  setProjectMinterConfigExtraMinterDetailsValue(
+    "approximateDAExpEndTime",
+    event.params.auctionTimestampStart.plus(totalAuctionTime),
     projectMinterConfig
   );
