+++ src/max-invocations-lib-mapping.ts
@@ -1,0 +1,41 @@
+import { BigInt, log, Address } from "@graphprotocol/graph-ts";
+import { Project } from "../generated/schema";
+
+import { ProjectMaxInvocationsLimitUpdated } from "../generated/MaxInvocationsLib/MaxInvocationsLib";
+
+import { updateProjectIfMinterConfigIsActive } from "./helpers";
+
+import { loadOrCreateMinterProjectAndConfigIfProject } from "./generic-minter-events-lib-mapping";
+
+///////////////////////////////////////////////////////////////////////////////
+// EVENT HANDLERS start here
+///////////////////////////////////////////////////////////////////////////////
+
+export function handleProjectMaxInvocationsLimitUpdated(
+  event: ProjectMaxInvocationsLimitUpdated
+): void {
+  const minterProjectAndConfig = loadOrCreateMinterProjectAndConfigIfProject(
+    event.address, // minter
+    event.params.coreContract,
+    event.params.projectId,
+    event.block.timestamp
+  );
+  if (!minterProjectAndConfig) {
+    // project wasn't found, warning already logged in helper function
+    return;
+  }
+  const projectMinterConfig = minterProjectAndConfig.projectMinterConfiguration;
+  projectMinterConfig.maxInvocations = event.params.maxInvocations;
+  projectMinterConfig.save();
+
+  // induce sync if the project minter configuration is the active one
+  updateProjectIfMinterConfigIsActive(
+    minterProjectAndConfig.project,
+    projectMinterConfig,
+    event.block.timestamp
+  );
+}
+
+///////////////////////////////////////////////////////////////////////////////
+// EVENT HANDLERS end here
+///////////////////////////////////////////////////////////////////////////////
