+++ src/merkle-lib-mapping.ts
@@ -1,7 +1,7 @@
 import {
   DefaultMaxInvocationsPerAddress,
   DelegationRegistryUpdated
-} from "../generated/ISharedMerkle/ISharedMinterMerkleV0";
+} from "../generated/MerkleLib/MerkleLib";
 
 import { loadOrCreateMinter } from "./helpers";
 
@@ -12,7 +12,7 @@ import { setMinterExtraMinterDetailsValue } from "./extra-minter-details-helpers
 ///////////////////////////////////////////////////////////////////////////////
 
 /**
- * Handles the event that updates the default max invocations per address. 
+ * Handles the event that updates the default max invocations per address.
  * Loads or creates the minter, updates the minter's details, and refreshes
  * the minter's timestamp to induce a sync.
  * @param event The event containing the updated max invocations per address.
@@ -36,7 +36,7 @@ export function handleDefaultMaxInvocationsPerAddress(
 }
 
 /**
- * Handles the event that updates the delegation registry address. 
+ * Handles the event that updates the delegation registry address.
  * Loads or creates the minter, updates the minter's details, and refreshes
  * the minter's timestamp to induce a sync.
  * @param event The event containing the updated delegation registry address.
