+++ src/split-funds-lib-mapping.ts
@@ -1,0 +1,48 @@
+import { BigInt, log, Address } from "@graphprotocol/graph-ts";
+import { Project } from "../generated/schema";
+
+import { ProjectCurrencyInfoUpdated } from "../generated/SplitFundsLib/SplitFundsLib";
+
+import { updateProjectIfMinterConfigIsActive } from "./helpers";
+
+import { loadOrCreateMinterProjectAndConfigIfProject } from "./generic-minter-events-lib-mapping";
+
+///////////////////////////////////////////////////////////////////////////////
+// EVENT HANDLERS start here
+///////////////////////////////////////////////////////////////////////////////
+
+/**
+ * Handles the update of a project's currency information. Attempts to load associated
+ * project and its minter configuration, then updates currency address and symbol.
+ * @param event The event carrying updated currency information
+ */
+export function handleProjectCurrencyInfoUpdated(
+  event: ProjectCurrencyInfoUpdated
+): void {
+  const minterProjectAndConfig = loadOrCreateMinterProjectAndConfigIfProject(
+    event.address, // minter
+    event.params.coreContract,
+    event.params.projectId,
+    event.block.timestamp
+  );
+  if (!minterProjectAndConfig) {
+    // project wasn't found, warning already logged in helper function
+    return;
+  }
+
+  const projectMinterConfig = minterProjectAndConfig.projectMinterConfiguration;
+  projectMinterConfig.currencyAddress = event.params.currencyAddress;
+  projectMinterConfig.currencySymbol = event.params.currencySymbol;
+  projectMinterConfig.save();
+
+  // induce sync if the project minter configuration is the active one
+  updateProjectIfMinterConfigIsActive(
+    minterProjectAndConfig.project,
+    projectMinterConfig,
+    event.block.timestamp
+  );
+}
+
+///////////////////////////////////////////////////////////////////////////////
+// EVENT HANDLERS end here
+///////////////////////////////////////////////////////////////////////////////
